working with browser login page now

+++ internal/model/auth.go
@@ -8,7 +8,7 @@ import (
 )
 
 type AuthClaims struct {
-	AccountId int    `json:"accountId"`
+	AccountId int64    `json:"accountId"`
 	User      string `json:"user"`
 	PwdHash   string `json:"pwdHash"`
